requirements.txt
# BAP Ontology Tools Dependencies

# YAML parsing (ruamel preserves formatting)
PyYAML>=6.0
ruamel.yaml>=0.18

# JSON Schema validation
jsonschema>=4.0

# Database connectivity (for sync)
psycopg2-binary>=2.9

# AI processing (FREE with Groq)
groq>=0.4
